requirements.txt
spacy==3.7.2
transformers==4.37.2
flask==2.0.2
flask-cors==3.0.10
twilio==7.1.0
google-api-python-client==2.26.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
chromadb==0.3.26
pandas==2.2.0
pydantic==1.10.12
werkzeug==2.0.3
python-dotenv
sendgrid
openai
# pip install torch torchvision torchaudio
# pip install tensorflow

# pip install spacy==3.3.0 thinc==8.0.17 pydantic==1.8.2 typer==0.4.1
# spacy==chromadb==0.3.26 pydantic==1.10.12